print list elements v1

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,3 +1,6 @@
+#include <stdlib.h>
+#include <string.h>
+#include <stdio.h>
 #include "lists.h"
 
 /**
@@ -8,7 +11,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 
 	while (h)
